social_media_api/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-filter==25.2
django-formtools==2.5.1
django-otp==1.6.3
django-phonenumber-field==8.3.0
django-taggit==6.1.0
django-two-factor-auth==1.18.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.95.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.18.0
mcp-mysql-tool==0.3.0
mysql-connector==2.2.9
mysql-connector-python==9.4.0
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
python3-openid==3.2.0
pywin32==311
qrcode==8.2
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
sniffio==1.3.1
social-auth-app-django==5.6.0
social-auth-core==4.8.1
sqlparse==0.5.3
sse-starlette==3.0.2
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uvicorn==0.38.0
whitenoise==6.11.0
